Revert "クラス化変更前"
This reverts commit 63e2c52b8818cfb34673fa74b6f2065dbf9db14f.

+++ DX 03/debug_font.cpp
@@ -58,7 +58,7 @@ void DebugFont_Finalize(void)
 void DebugFont_Draw(int x, int y, const char* pFormat, ...)
 {
 #if defined(_DEBUG) || defined(DEBUG)
-	RECT rect = { (long)x, (long)y, (long)SCREEN_WIDTH, (long)SCREEN_HEIGHT 
+	RECT rect = { x, y, SCREEN_WIDTH, SCREEN_HEIGHT 
 	};
 	
 	va_list argp;
